refactor(characters): extract Characters.Leveling
Third slice of #134. XP threshold derivation, xp_to_next_level, and
at-level DAG evaluation move together as the level-math concern.
level_xp_thresholds, xp_effect_target, and evaluate_at_level were
private but are needed by compute_pending_choice_slots (which will land
in Characters.Progression), so they become public with @doc false —
internal API between the concern modules, hidden from Hex docs.
