+++ apps/ex_ttrpg_dev/lib/characters.ex
@@ -8,12 +8,11 @@ defmodule ExTTRPGDev.Characters do
   alias ExTTRPGDev.Characters.Decision
   alias ExTTRPGDev.Characters.Effects
   alias ExTTRPGDev.Characters.InventoryItem
+  alias ExTTRPGDev.Characters.Leveling
   alias ExTTRPGDev.Characters.Store
-  alias ExTTRPGDev.RuleSystem.Evaluator
   alias ExTTRPGDev.RuleSystem.Expression
   alias ExTTRPGDev.RuleSystem.Effect
   alias ExTTRPGDev.RuleSystem.InventoryRules
-  alias ExTTRPGDev.RuleSystem.Node
   alias ExTTRPGDev.RuleSystem.Vocabulary
   alias ExTTRPGDev.RuleSystems.LoadedSystem
 
@@ -120,42 +119,7 @@ defmodule ExTTRPGDev.Characters do
 
   Returns `{:error, :no_level_thresholds}` if the system does not define a level mapping.
   """
-  def xp_to_next_level(%LoadedSystem{} = system, %Character{} = character) do
-    thresholds = level_xp_thresholds(system)
-
-    if map_size(thresholds) == 0 do
-      {:error, :no_level_thresholds}
-    else
-      xp_target = xp_effect_target(system)
-
-      current_xp =
-        character.effects
-        |> Enum.filter(&(&1.target == xp_target))
-        |> Enum.map(& &1.value)
-        |> Enum.sum()
-
-      current_level =
-        thresholds
-        |> Enum.filter(fn {_level, threshold} -> threshold <= current_xp end)
-        |> Enum.max_by(fn {_level, threshold} -> threshold end, fn -> nil end)
-        |> case do
-          nil -> 1
-          {level, _} -> level
-        end
-
-      next_level =
-        thresholds
-        |> Map.keys()
-        |> Enum.sort()
-        |> Enum.find(&(&1 > current_level))
-
-      if is_nil(next_level) do
-        {:error, :max_level}
-      else
-        {:ok, Map.fetch!(thresholds, next_level) - current_xp, next_level}
-      end
-    end
-  end
+  defdelegate xp_to_next_level(system, character), to: Leveling
 
   @doc """
   Applies an award concept to a character.
@@ -752,8 +716,8 @@ defmodule ExTTRPGDev.Characters do
       {all_effects, resolved} = resolved_state(system, character)
       current_level = trunc(resolved[level_node_key] || 1)
 
-      thresholds = level_xp_thresholds(system)
-      xp_target = xp_effect_target(system)
+      thresholds = Leveling.level_xp_thresholds(system)
+      xp_target = Leveling.xp_effect_target(system)
 
       selection_progressions =
         Enum.filter(system.concept_metadata, fn {{type, _id}, meta} ->
@@ -764,7 +728,15 @@ defmodule ExTTRPGDev.Characters do
 
       level_resolved =
         Map.new(1..current_level, fn level ->
-          {level, evaluate_at_level(system, character, level, thresholds, xp_target, all_effects)}
+          {level,
+           Leveling.evaluate_at_level(
+             system,
+             character,
+             level,
+             thresholds,
+             xp_target,
+             all_effects
+           )}
         end)
 
       Enum.flat_map(selection_progressions, fn {{_type, id}, meta} ->
@@ -1127,50 +1099,6 @@ defmodule ExTTRPGDev.Characters do
     |> Enum.flat_map(fn {_, choice_def} ->
       if choice_def["type"] == type_id, do: choice_def["options"] || [], else: []
     end)
-  end
-
-  defp level_xp_thresholds(%LoadedSystem{} = system) do
-    with level_node when not is_nil(level_node) <- system.module.level_node,
-         [{type_id, concept_id, field_name} | _] <- Expression.extract_refs(level_node),
-         %Node{type: :mapping, steps: steps} when not is_nil(steps) <-
-           Map.get(system.nodes, {type_id, concept_id, field_name}) do
-      Map.new(steps, fn [threshold, level] -> {level, threshold} end)
-    else
-      _ -> %{}
-    end
-  end
-
-  defp xp_effect_target(%LoadedSystem{} = system) do
-    with level_node when not is_nil(level_node) <- system.module.level_node,
-         [{type_id, concept_id, field_name} | _] <- Expression.extract_refs(level_node),
-         %Node{type: :mapping, input: input} when not is_nil(input) <-
-           Map.get(system.nodes, {type_id, concept_id, field_name}),
-         [node_key | _] <- Expression.extract_refs(input) do
-      node_key
-    else
-      _ -> nil
-    end
-  end
-
-  defp evaluate_at_level(
-         %LoadedSystem{} = system,
-         %Character{} = character,
-         level,
-         thresholds,
-         xp_target,
-         all_effects
-       ) do
-    xp_for_level = Map.get(thresholds, level, 0)
-    non_xp_effects = Enum.reject(all_effects, &(&1.target == xp_target))
-
-    level_effects =
-      if xp_target && xp_for_level > 0 do
-        [%Effect{target: xp_target, value: xp_for_level} | non_xp_effects]
-      else
-        non_xp_effects
-      end
-
-    Evaluator.evaluate!(system, character.generated_values, level_effects)
   end
 
   # --- activate/4 helpers ---
